chore: connecting BD and creating enviroment variables to use with

+++ backend/src/main/java/com/ufrn/projeto/webApp/dto/DashboardDTO.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -21,7 +22,7 @@ public class DashboardDTO {
     private String nomePet;
     private String racaPet;
     private Porte portePet;
-    private Double pesoPet;
+    private BigDecimal pesoPet;
     private Sexo sexoPet;
 
 }
